+++ CHANGELOG.md
@@ -54,3 +54,7 @@ Added Auth Layouts
     <option></option>
 </x-wb::select>
 ```
+
+## v1.3.0 Added Modal & Dialog component - 2022-04-20
+
+-   Added `<x-wb::dialog></x-wb::dialog>` component which is same as available in Laravel Jetstream